chore(): rename base-chart to chart
