+++ src/platform/echarts/base/axis/axis.component.ts
@@ -5,7 +5,7 @@ import {
   OnDestroy,
 } from '@angular/core';
 
-import { TdChartOptionsService } from '../base.service';
+import { TdChartOptionsService } from '../chart.service';
 import { assignDefined } from '../utils';
 
 export abstract class TdChartAxisComponent implements OnChanges, OnInit, OnDestroy {
