+++ src/platform/echarts/base/axis/x-axis.component.ts
@@ -4,7 +4,7 @@ import {
   Input,
 } from '@angular/core';
 
-import { TdChartOptionsService } from '../base.service';
+import { TdChartOptionsService } from '../chart.service';
 import { TdChartAxisComponent } from './axis.component';
 
 @Component({
