+++ src/platform/echarts/base/base.module.ts
@@ -1,13 +1,13 @@
 import { NgModule, Type } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { TdBaseChartComponent } from './base.component';
+import { TdChartComponent } from './chart.component';
 import { TdChartTooltipComponent, TdChartTooltipFormatterDirective } from './tooltip/tooltip.component';
 import { TdChartXAxisComponent } from './axis/x-axis.component';
 import { TdChartYAxisComponent } from './axis/y-axis.component';
 
 export const BASE_MODULE_COMPONENTS: Type<any>[] = [
-  TdBaseChartComponent,
+  TdChartComponent,
   TdChartTooltipComponent,
   TdChartTooltipFormatterDirective,
   TdChartXAxisComponent,
