+++ src/platform/echarts/base/chart.component.scss
