+++ src/platform/echarts/base/public-api.ts
@@ -1,3 +1,3 @@
 export * from './base.module';
-export { TdBaseChartComponent } from './base.component';
-export { TdChartOptionsService, BASE_CHART_PROVIDER } from './base.service';
+export { TdChartComponent } from './chart.component';
+export { TdChartOptionsService, CHART_PROVIDER } from './chart.service';
